[refactor]: Rename lib -> utils
