+++ src/main/java/lotto/utils/validator/PriceValidator.java
@@ -1,4 +1,4 @@
-package lotto.lib.validator;
+package lotto.utils.validator;
 
 import lotto.exception.BadPriceException;
