+++ src/test/java/lotto/utils/validator/LottoValidatorTest.java
@@ -1,4 +1,4 @@
-package lotto.lib.validator;
+package lotto.utils.validator;
 
 import lotto.asset.ExceptionConst;
 import lotto.exception.BadNumOfLottoNoException;
